final_artifacts/01_OBJETO_DE_ESTUDO_AMBIENTE/docker_testbed/requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
cicflowmeter==0.5.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
idna==3.13
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.9
numpy==2.4.4
packaging==26.2
paho-mqtt==2.1.0
pandas==3.0.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.33.1
scapy==2.7.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
threadpoolctl==3.6.0
urllib3==2.6.3
